backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
motor>=3.3
python-dotenv>=1.0
pydantic>=2.5
starlette>=0.37
